Use Bigint for userid in feds_settings
query to migrate in db

ALTER TABLE feds_settings ALTER COLUMN user_id SET DATA TYPE BigInt;

+++ tg_bot/modules/sql/feds_sql.py
@@ -1,6 +1,7 @@
 import threading
 import ast
 from sqlalchemy import Column, String, UnicodeText, Integer, Boolean
+from sqlalchemy.sql.sqltypes import BigInteger
 from telegram.error import BadRequest, Unauthorized
 
 from tg_bot import dispatcher
@@ -59,7 +60,7 @@ class BansF(BASE):
 
 class FedsUserSettings(BASE):
     __tablename__ = "feds_settings"
-    user_id = Column(Integer, primary_key=True)
+    user_id = Column(BigInteger, primary_key=True)
     should_report = Column(Boolean, default=True)
 
     def __init__(self, user_id):
